requirements.txt
aiofiles>=25.1.0
chromadb>=1.5.0
fastapi>=0.128.8
langchain-community>=0.4.1
langchain-core>=1.2.11
langchain-google-genai>=1.0.0
pandas>=2.3.3
passlib>=1.7.4
plotly>=6.5.2
psycopg2-binary>=2.9.11
pypdf2>=3.0.1
python-dotenv>=1.2.1
python-jose[cryptography]>=3.5.0
python-multipart>=0.0.22
requests>=2.32.5
sqlalchemy>=2.0.46
streamlit>=1.54.0
uvicorn>=0.40.0
uv>= 0.10.2